fix(build): resolve compilation errors related to scoring overhaul

+++ ailab-super-app/DTOs/Statistics/LeaderboardUserDto.cs
@@ -3,7 +3,7 @@ namespace ailab_super_app.DTOs.Statistics
     public class LeaderboardUserDto
     {
         public string FullName { get; set; } = string.Empty;
-        public int TotalScore { get; set; }
+        public decimal TotalScore { get; set; }
         public string? ProfileImageUrl { get; set; }
     }
 }
